+++ AndroidTestsProject/app/src/androidTest/java/com/distrimind/flexilogxml/android/TestNGRunner.java
@@ -50,7 +50,6 @@ public class TestNGRunner {
 
 	@Test
 	public void allTestNG() throws Throwable {
-		Log.w("testDirectAndroidLog", new IllegalAccessError());
 		tests.runTestNGWithJunit("TestUtils");
 	}
 }