+++ changelog.md
@@ -10,14 +10,17 @@ DM-FlexiLogXML
 
 # Changes:
 
-### 1.2.3 STABLE (2025-01-28)
+### 1.2.8 STABLE (2025-01-28)
+#### New feature(s)
+* Add standalone parameter into function writeStartDocument of interface IWXmlWriter
 #### Fix bug
 * Fix issue with attribute writing under Android
+* Fix issue with end tag writing under Android
+* Use the same logger factory
+* Better filter logs through log levels
 
 
 ### 1.2.3 STABLE (2025-01-28)
-#### New feature(s)
-* Add standalone parameter into function writeStartDocument of interface IWXmlWriter  
 
 ### 1.0.0 STABLE (2025-01-15)
 #### New feature(s)